# api-docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "Backend API For the Auth/User Service",
    "version": "Unreleased [WIP]",
    "title": "Auth/User API"
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "paths": {
    "/auth/refresh": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Refresh the expired jwt authentication by issuing a token refresh request and returns theupdated response tokens",
        "operationId": "refreshJwtTokenUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "tokenRefreshRequest",
            "description": "The TokenRefreshRequest payload",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Token refresh Request"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/JWT Authentication Response"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/auth/signin": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "Logs the user in to the system and return the auth tokens and some other data",
        "operationId": "signinUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "signinRequest",
            "description": "signinRequest",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Signin Request DTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Authentication Response"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/auth/signup": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Register new user, then log them in. If succeed return the auth tokens and some other data",
        "operationId": "signupUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "signupRequest",
            "description": "signupRequest",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Signup Request DTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Authentication Response"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/in_queue": {
      "get": {
        "tags": [
          "Manage Queue",
          "Manage User in Queue"
        ],
        "summary": "Returns all users/ clients that's currently inside the specified queue.This endpoint can only be hit by a queue owner.",
        "operationId": "inQueuesUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "inQueueRequest",
            "description": "inQueueRequest",
            "required": true,
            "schema": {
              "$ref": "#/definitions/InQueue Request DTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/In Queue Response DTO"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Manage Queue",
          "Manage User in Queue"
        ],
        "summary": "Add a client to a queue, if succeed will return newly created client with userId set to -1Lbecause no registered user is associated with it. Because of that client added to a queue this way will notbe notify about queue changes via websocket. This endpoint can only be hit by a queue owner.",
        "operationId": "addToInQueueUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "joinQueueRequest",
            "description": "joinQueueRequest",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Join Queue Request DTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/In Queue Response DTO"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Manage Queue",
          "Manage User in Queue"
        ],
        "summary": "Remove an unwanted user from a queue.If succeed or no record found with the specified id, will return success message.This endpoint can only be hit by a queue owner.",
        "operationId": "remoteFromInQueueUsingDELETE",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "removeFromInQueueRequest",
            "description": "removeFromInQueueRequest",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Remove From In Queue Request DTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Common API Response"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
...
}